requirements.txt
python-dotenv

python-pptx

rich
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-mcp-adapters>=0.0.1
langgraph>=0.0.15

# "mcp[cli]"
fastmcp